requirements.txt
streamlit
python-dotenv
groq
spotipy